Moved Get Handler.
